// src/Api/ClientApp/src/api/clientApi.ts
export interface TranslationJobDto {
    id: number;
    customerName?: string | undefined;
    status: JobStatus;
    originalContent?: string | undefined;
    translatedContent?: string | undefined;
    price: number;
    translatorId?: number | undefined;
}

export enum JobStatus {
    New = 1,
    InProgress = 2,
    Completed = 3,
}

export interface TranslationJobDetailDto extends TranslationJobDto {
    translator?: TranslatorDto | undefined;
}

export interface TranslatorDto {
    id: number;
    name?: string | undefined;
    hourlyRate: number;
    status: TranslatorStatus;
}

export enum TranslatorStatus {
    Applicant = 1,
    Certified = 2,
    Deleted = 3,
}

export interface TranslatorDetailDto extends TranslatorDto {
    translationJobs?: TranslationJobRefDto[] | undefined;
}

export interface TranslationJobRefDto {
    id: number;
    customerName?: string | undefined;
    status: JobStatus;
    price: number;
}